refactor: move Extractions tab to Test
