+++ frontend/app/v2/results/page.tsx
@@ -3,11 +3,9 @@
 import { useState, useEffect, useCallback, useMemo, useRef } from 'react'
 import { useSearchParams, useRouter } from 'next/navigation'
 import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import { Badge } from '@/components/ui/badge'
+import { Card, CardContent } from '@/components/ui/card'
 
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
-import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible'
 import { PapersTable } from '@/components/search/papers-table'
 import { InterventionsTable as InterventionsEvidenceTable, InterventionData } from '@/components/search/interventions-table'
 import { Paper } from '@/types/search'
@@ -17,10 +15,7 @@ import {
   ArrowLeft,
   AlertCircle,
   BookOpen,
-  ChevronRight,
-  ChevronDown,
   Target,
-  TrendingUp,
   Bot
 } from 'lucide-react'
 import { useAnalysisProjectStore } from '@/lib/analysisProjectStore'
@@ -54,313 +49,7 @@ interface AnalysisDocument {
   cited_by_count?: number  // Citation count from API
 }
 
-interface DocumentDetailResult {
-  document: {
-    id: string
-    doc_id: string
-    title: string
-    source: string
-    year?: number
-    abstract_or_summary?: string
-    is_relevant?: boolean
-    extraction_status?: string
-  }
-  extraction: {
-    issues: Array<{
-      idx?: number
-      label?: string
-      explanation?: string
-      supporting_quote?: string
-    }>
-    interventions: Array<{
-      idx?: number
-      name?: string
-      description?: string
-      type?: string
-      country?: string
-      study_type?: string
-      supporting_quote?: string
-      addresses_issues?: number[]
-      results?: Array<{
-        outcome_variable?: string
-        effect_direction?: string
-        effect_size_type?: string
-        effect_size?: string
-        uncertainty?: string
-        p_value?: string
-        population_measured?: string
-        subgroup_or_dose?: string
-        result_text?: string
-        supporting_quote?: string
-      }>
-    }>
-    mappings?: unknown[]
-    conclusion?: {
-      top_line_summary?: string
-      detailed_explanation?: string
-      supporting_quote?: string
-    }
-    metadata?: Record<string, unknown>
-  }
-}
 
-// Document Detail View Component
-function DocumentDetailView({ extraction }: { 
-  extraction: DocumentDetailResult['extraction']
-}) {
-  const [openSections, setOpenSections] = useState({
-    issues: true,
-    interventions: true,
-    results: false,
-    conclusion: true
-  })
-
-  const toggleSection = (section: keyof typeof openSections) => {
-    setOpenSections(prev => ({ ...prev, [section]: !prev[section] }))
-  }
-
-  const issues = extraction.issues || []
-  const interventions = extraction.interventions || []
-  const conclusion = extraction.conclusion
-
-  return (
-    <div className="space-y-4">
-      {/* Document Header */}
-
-
-      {/* Issues Section */}
-      <Collapsible open={openSections.issues} onOpenChange={() => toggleSection('issues')}>
-        <CollapsibleTrigger asChild>
-          <Card className="cursor-pointer hover:bg-gray-50 border-red-200">
-            <CardHeader className="pb-3">
-              <CardTitle className="flex items-center justify-between text-base">
-                <div className="flex items-center gap-2">
-                  <AlertCircle className="h-4 w-4 text-red-600" />
-                  <span className="text-red-900">Issues & Problems ({issues.length})</span>
-                </div>
-                {openSections.issues ? 
-                  <ChevronDown className="h-4 w-4 text-gray-500" /> : 
-                  <ChevronRight className="h-4 w-4 text-gray-500" />
-                }
-              </CardTitle>
-            </CardHeader>
-          </Card>
-        </CollapsibleTrigger>
-        <CollapsibleContent>
-          <div className="space-y-3 mt-2">
-            {issues.map((issue, index: number) => (
-              <Card key={issue.idx || index} className="border-red-100">
-                <CardContent className="p-4">
-                  <div className="flex items-start gap-3">
-                    <div className="w-6 h-6 bg-red-100 rounded-full flex items-center justify-center flex-shrink-0">
-                                              <span className="text-red-700 font-medium text-sm">{issue.idx !== undefined ? issue.idx : index + 1}</span>
-                    </div>
-                    <div className="flex-1">
-                      <h5 className="font-medium text-red-900 mb-1">{issue.label}</h5>
-                      <p className="text-red-700 text-sm mb-2">{issue.explanation}</p>
-                      {issue.supporting_quote && (
-                        <blockquote className="text-red-600 text-xs italic border-l-4 border-red-200 pl-3">
-                          &ldquo;{issue.supporting_quote}&rdquo;
-                        </blockquote>
-                      )}
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            ))}
-            {issues.length === 0 && (
-              <p className="text-gray-500 text-sm italic">No issues identified in this document.</p>
-            )}
-          </div>
-        </CollapsibleContent>
-      </Collapsible>
-
-      {/* Interventions & Results Section */}
-      <Collapsible open={openSections.interventions} onOpenChange={() => toggleSection('interventions')}>
-        <CollapsibleTrigger asChild>
-          <Card className="cursor-pointer hover:bg-gray-50 border-blue-200">
-            <CardHeader className="pb-3">
-              <CardTitle className="flex items-center justify-between text-base">
-                <div className="flex items-center gap-2">
-                  <Target className="h-4 w-4 text-blue-600" />
-                  <span className="text-blue-900">Interventions & Results ({interventions.length})</span>
-                </div>
-                {openSections.interventions ? 
-                  <ChevronDown className="h-4 w-4 text-gray-500" /> : 
-                  <ChevronRight className="h-4 w-4 text-gray-500" />
-                }
-              </CardTitle>
-            </CardHeader>
-          </Card>
-        </CollapsibleTrigger>
-        <CollapsibleContent>
-          <div className="space-y-4 mt-2">
-            {interventions.map((intervention, index: number) => (
-              <Card key={intervention.idx || index} className="border-blue-100">
-                <CardContent className="p-4">
-                  <div className="space-y-3">
-                    {/* Intervention Header */}
-                    <div className="flex items-start gap-3">
-                      <div className="w-6 h-6 bg-blue-100 rounded-full flex items-center justify-center flex-shrink-0">
-                        <span className="text-blue-700 font-medium text-sm">{intervention.idx !== undefined ? intervention.idx : index + 1}</span>
-                      </div>
-                      <div className="flex-1">
-                        <h5 className="font-medium text-blue-900 mb-1">{intervention.name}</h5>
-                        <p className="text-blue-700 text-sm mb-2">{intervention.description}</p>
-                        
-                        {/* Intervention Details */}
-                        <div className="flex gap-2 mb-2">
-                          <Badge variant="outline" className="text-xs bg-blue-50">
-                            {intervention.type}
-                          </Badge>
-                          {intervention.country && (
-                            <Badge variant="outline" className="text-xs bg-blue-50">
-                              {intervention.country}
-                            </Badge>
-                          )}
-                          {intervention.study_type && (
-                            <Badge variant="outline" className="text-xs bg-blue-50">
-                              Study: {intervention.study_type}
-                            </Badge>
-                          )}
-                        </div>
-
-                        {/* Issues Addressed */}
-                        {intervention.addresses_issues && intervention.addresses_issues.length > 0 && (
-                          <div className="mb-2">
-                            <span className="text-xs text-blue-600 font-medium">Addresses issues: </span>
-                            {intervention.addresses_issues.map((issueIdx: number) => {
-                              const relatedIssue = issues.find(issue => issue.idx === issueIdx)
-                              const issueTitle = relatedIssue?.label || `Issue ${issueIdx}`
-                              return (
-                                <Badge key={issueIdx} variant="outline" className="text-xs mr-1 bg-red-50 text-red-700" title={issueTitle}>
-                                  #{issueIdx}: {issueTitle}
-                                </Badge>
-                              )
-                            })}
-                          </div>
-                        )}
-
-                        {/* Results for this intervention */}
-                        {intervention.results && intervention.results.length > 0 && (
-                          <div className="mt-3">
-                            <h6 className="font-medium text-green-900 text-sm mb-2 flex items-center gap-1">
-                              <TrendingUp className="h-3 w-3" />
-                              Results ({intervention.results.length})
-                            </h6>
-                            <div className="space-y-2">
-                              {intervention.results.map((result, resultIndex: number) => (
-                                <div key={resultIndex} className="bg-green-50 border-l-4 border-green-200 p-2 rounded">
-                                  <div className="flex items-center gap-2 mb-1">
-                                    <span className="font-medium text-green-900 text-sm">
-                                      {result.outcome_variable}
-                                    </span>
-                                    <Badge variant="outline" className="text-xs bg-green-100 text-green-700">
-                                      {result.effect_direction}
-                                    </Badge>
-                                  </div>
-                                  
-                                  {/* Quantitative measures */}
-                                  {(result.effect_size || result.effect_size_type || result.uncertainty || result.p_value) && (
-                                    <div className="mb-2">
-                                      {result.effect_size_type && (
-                                        <div className="text-xs text-green-600 mb-1">
-                                          <span className="font-medium">Effect Type: </span>
-                                          {result.effect_size_type}
-                                        </div>
-                                      )}
-                                      {result.effect_size && (
-                                        <div className="text-xs text-green-600 mb-1">
-                                          <span className="font-medium">Effect Size: </span>
-                                          {result.effect_size}
-                                        </div>
-                                      )}
-                                      {result.uncertainty && (
-                                        <div className="text-xs text-green-600 mb-1">
-                                          <span className="font-medium">Uncertainty: </span>
-                                          ±{result.uncertainty}
-                                        </div>
-                                      )}
-                                      {result.p_value && (
-                                        <div className="text-xs text-green-600 mb-1">
-                                          <span className="font-medium">P-value: </span>
-                                          {result.p_value}
-                                        </div>
-                                      )}
-                                    </div>
-                                  )}
-                                  
-                                  {/* Population information */}
-                                  {result.population_measured && (
-                                    <div className="mb-2">
-                                      <div className="text-xs text-green-600 mb-1">
-                                        <span className="font-medium">Population: </span>
-                                        {result.population_measured}
-                                      </div>
-                                    </div>
-                                  )}
-                                  
-                                  <p className="text-green-700 text-xs">{result.result_text}</p>
-                                  {result.supporting_quote && (
-                                    <blockquote className="text-green-600 text-xs italic mt-1">
-                                      &ldquo;{result.supporting_quote}&rdquo;
-                                    </blockquote>
-                                  )}
-                                </div>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            ))}
-            {interventions.length === 0 && (
-              <p className="text-gray-500 text-sm italic">No interventions identified in this document.</p>
-            )}
-          </div>
-        </CollapsibleContent>
-      </Collapsible>
-
-      {/* Conclusion Section */}
-      {conclusion && (
-        <Collapsible open={openSections.conclusion} onOpenChange={() => toggleSection('conclusion')}>
-          <CollapsibleTrigger asChild>
-            <Card className="cursor-pointer hover:bg-gray-50 border-purple-200">
-              <CardHeader className="pb-3">
-                <CardTitle className="flex items-center justify-between text-base">
-                  <div className="flex items-center gap-2">
-                    <FileText className="h-4 w-4 text-purple-600" />
-                    <span className="text-purple-900">Conclusion</span>
-                  </div>
-                  {openSections.conclusion ? 
-                    <ChevronDown className="h-4 w-4 text-gray-500" /> : 
-                    <ChevronRight className="h-4 w-4 text-gray-500" />
-                  }
-                </CardTitle>
-              </CardHeader>
-            </Card>
-          </CollapsibleTrigger>
-          <CollapsibleContent>
-            <Card className="border-purple-100 mt-2">
-              <CardContent className="p-4">
-                <h5 className="font-medium text-purple-900 mb-2">{conclusion.top_line_summary}</h5>
-                <p className="text-purple-700 text-sm mb-3">{conclusion.detailed_explanation}</p>
-                {conclusion.supporting_quote && (
-                  <blockquote className="text-purple-600 text-xs italic border-l-4 border-purple-200 pl-3">
-                    &ldquo;{conclusion.supporting_quote}&rdquo;
-                  </blockquote>
-                )}
-              </CardContent>
-            </Card>
-          </CollapsibleContent>
-        </Collapsible>
-      )}
-    </div>
-  )
-}
 
 export default function AnalysisResultsPage() {
   const searchParams = useSearchParams()
@@ -384,15 +73,9 @@ export default function AnalysisResultsPage() {
   const [interventions, setInterventions] = useState<InterventionData[]>([])
   const [loadingData, setLoadingData] = useState(false)
   const [dataError, setDataError] = useState<string | null>(null)
-  
-  // Document detail view state
-  const [selectedDocumentId, setSelectedDocumentId] = useState<string | null>(null)
-  const [documentDetail, setDocumentDetail] = useState<DocumentDetailResult | null>(null)
-  const [loadingDocumentDetail, setLoadingDocumentDetail] = useState(false)
-  const [documentDetailError, setDocumentDetailError] = useState<string | null>(null)
 
   const { activeProject } = useAnalysisProjectStore()
-  const { fetchWithAuth, getDocumentExtraction, getAnalysisProject, getProjectInterventions } = useAPI()
+  const { fetchWithAuth, getAnalysisProject, getProjectInterventions } = useAPI()
 
   // Get search parameters (memoized to prevent re-runs)
   const searchConfig = useMemo(() => ({
@@ -729,31 +412,6 @@ export default function AnalysisResultsPage() {
     router.push('/v2/search')
   }
 
-  // Load document detail extraction
-  const loadDocumentDetail = useCallback(async (documentId: string) => {
-    if (!effectiveProjectId) return
-    
-    setLoadingDocumentDetail(true)
-    setDocumentDetailError(null)
-    setSelectedDocumentId(documentId)
-    
-    try {
-      const result = await getDocumentExtraction(effectiveProjectId, documentId)
-      setDocumentDetail(result)
-    } catch (error) {
-      console.error('Failed to load document detail:', error)
-      setDocumentDetailError(error instanceof Error ? error.message : 'Failed to load document detail')
-    } finally {
-      setLoadingDocumentDetail(false)
-    }
-  }, [effectiveProjectId, getDocumentExtraction])
-
-  // Close document detail view
-  const closeDocumentDetail = () => {
-    setSelectedDocumentId(null)
-    setDocumentDetail(null)
-    setDocumentDetailError(null)
-  }
 
   // Transform documents for table display
   const transformedPapers: Paper[] = documents.map((doc: AnalysisDocument) => ({
@@ -869,7 +527,7 @@ export default function AnalysisResultsPage() {
         {effectiveProjectId && (
           <Tabs value={activeTab} onValueChange={setActiveTab} className="h-full flex flex-col">
             <div className="px-6 pt-4">
-              <TabsList className="grid w-full grid-cols-4">
+              <TabsList className="grid w-full grid-cols-3">
                 <TabsTrigger value="evidence" className="flex items-center gap-2">
                   <BookOpen className="h-4 w-4" />
                   Evidence
@@ -877,10 +535,6 @@ export default function AnalysisResultsPage() {
                 <TabsTrigger value="summary" className="flex items-center gap-2">
                   <FileText className="h-4 w-4" />
                   Summary
-                </TabsTrigger>
-                <TabsTrigger value="extraction" className="flex items-center gap-2">
-                  <FileText className="h-4 w-4" />
-                  Extraction
                 </TabsTrigger>
                 <TabsTrigger value="assistant" className="flex items-center gap-2">
                   <Bot className="h-4 w-4" />
@@ -890,168 +544,6 @@ export default function AnalysisResultsPage() {
             </div>
 
             <div className="flex-1 overflow-auto">
-              <TabsContent value="extraction" className="p-6 m-0">
-                <div className="max-w-6xl mx-auto">
-
-
-                  {/* Document List */}
-                  {loadingData ? (
-                    <div className="flex items-center justify-center py-12">
-                      <div className="text-center">
-                        <Loader2 className="h-8 w-8 animate-spin mx-auto mb-4" />
-                        <p className="text-slate-600">Loading documents...</p>
-                      </div>
-                    </div>
-                  ) : dataError ? (
-                    <div className="text-center py-12">
-                      <AlertCircle className="h-12 w-12 text-red-400 mx-auto mb-4" />
-                      <h3 className="text-lg font-medium text-slate-900 mb-2">Error Loading Data</h3>
-                      <p className="text-slate-600">{dataError}</p>
-                    </div>
-                  ) : documents.length > 0 ? (
-                    <div className="space-y-4">
-                      <div className="flex items-center justify-between">
-                      </div>
-                      <div className="space-y-3">
-                        {documents.map((doc) => {
-                          const isExpanded = selectedDocumentId === doc.id
-                          const handleCardClick = (e: React.MouseEvent) => {
-                            // Don't trigger if clicking on a link
-                            if ((e.target as HTMLElement).closest('a')) {
-                              return
-                            }
-                            
-                            if (isExpanded) {
-                              closeDocumentDetail()
-                            } else {
-                              loadDocumentDetail(doc.id)
-                            }
-                          }
-                          
-                          return (
-                            <Card 
-                              key={doc.id} 
-                              className={`border-slate-200 cursor-pointer transition-all hover:shadow-md ${
-                                isExpanded ? 'ring-2 ring-blue-300 bg-blue-50' : ''
-                              }`}
-                              onClick={handleCardClick}
-                            >
-                              <CardContent className="p-4">
-                                {/* Card Header - Always Visible */}
-                                <div className="flex items-start justify-between">
-                                  <div className="flex-1">
-                                    <h3 className="font-semibold text-slate-900 mb-2">
-                                      {doc.landing_page_url ? (
-                                        <a 
-                                          href={doc.landing_page_url} 
-                                          target="_blank" 
-                                          rel="noopener noreferrer"
-                                          className="text-blue-600 hover:text-blue-800 hover:underline"
-                                          onClick={(e) => e.stopPropagation()}
-                                        >
-                                          {doc.title || 'Untitled Document'}
-                                        </a>
-                                      ) : (
-                                        doc.title || 'Untitled Document'
-                                      )}
-                                    </h3>
-                                    <div className="flex items-center gap-2 mb-2">
-                                      <Badge variant="outline" className="text-xs">
-                                        {doc.source || 'Unknown Source'}
-                                      </Badge>
-                                      {doc.year && (
-                                        <Badge variant="outline" className="text-xs">
-                                          {doc.year}
-                                        </Badge>
-                                      )}
-                                      {doc.is_relevant !== null && (
-                                        <Badge 
-                                          variant="outline" 
-                                          className={`text-xs ${
-                                            doc.is_relevant 
-                                              ? 'bg-green-100 text-green-700' 
-                                              : 'bg-red-100 text-red-700'
-                                          }`}
-                                        >
-                                          {doc.is_relevant ? 'Relevant' : 'Not Relevant'}
-                                        </Badge>
-                                      )}
-                                      {/* Extraction Status Indicator */}
-                                      {doc.extraction_status && (
-                                        <Badge 
-                                          variant="outline" 
-                                          className={`text-xs ${
-                                            doc.extraction_status === 'success' 
-                                              ? 'bg-blue-100 text-blue-700' 
-                                              : doc.extraction_status === 'failed'
-                                              ? 'bg-red-100 text-red-700'
-                                              : doc.extraction_status === 'skipped'
-                                              ? 'bg-yellow-100 text-yellow-700'
-                                              : 'bg-gray-100 text-gray-700'
-                                          }`}
-                                        >
-                                          {doc.extraction_status === 'success' ? '✓ Extracted' :
-                                           doc.extraction_status === 'failed' ? '✗ Failed' :
-                                           doc.extraction_status === 'skipped' ? '⊘ Skipped' :
-                                           `✓ ${doc.extraction_status}`}
-                                        </Badge>
-                                      )}
-                                    </div>
-                                    {!isExpanded && doc.abstract_or_summary && (
-                                      <p className="text-slate-600 text-sm line-clamp-2 mb-2">
-                                        {doc.abstract_or_summary}
-                                      </p>
-                                    )}
-                                  </div>
-                                  <div className="flex items-center gap-2 ml-4">
-                                    {isExpanded ? (
-                                      <ChevronDown className="h-4 w-4 text-slate-400" />
-                                    ) : (
-                                      <ChevronRight className="h-4 w-4 text-slate-400" />
-                                    )}
-                                  </div>
-                                </div>
-                                
-                                {/* Expanded Content */}
-                                {isExpanded && (
-                                  <div className="mt-4 pt-4 border-t border-slate-200">
-                                    {loadingDocumentDetail ? (
-                                      <div className="flex items-center justify-center py-8">
-                                        <div className="text-center">
-                                          <Loader2 className="h-6 w-6 animate-spin mx-auto mb-2 text-blue-600" />
-                                          <p className="text-blue-700 text-sm">Loading document details...</p>
-                                        </div>
-                                      </div>
-                                    ) : documentDetailError ? (
-                                      <div className="text-center py-4">
-                                        <AlertCircle className="h-8 w-8 text-red-400 mx-auto mb-2" />
-                                        <p className="text-red-600 text-sm">{documentDetailError}</p>
-                                      </div>
-                                    ) : documentDetail && documentDetail.extraction ? (
-                                      <DocumentDetailView extraction={documentDetail.extraction} />
-                                    ) : documentDetail ? (
-                                      <div className="text-center py-4">
-                                        <FileText className="h-8 w-8 text-gray-400 mx-auto mb-2" />
-                                        <p className="text-gray-600 text-sm">No extraction data available for this document</p>
-                                      </div>
-                                    ) : null}
-                                  </div>
-                                )}
-                              </CardContent>
-                            </Card>
-                          )
-                        })}
-                      </div>
-                    </div>
-                  ) : (
-                    <div className="text-center py-12">
-                      <FileText className="h-12 w-12 text-slate-400 mx-auto mb-4" />
-                      <h3 className="text-lg font-medium text-slate-900 mb-2">No Documents Available</h3>
-                      <p className="text-slate-600">Documents will appear here once the analysis is processed.</p>
-                    </div>
-                  )}
-                </div>
-              </TabsContent>
 
               <TabsContent value="summary" className="p-6 m-0">
                 <div className="max-w-6xl mx-auto">
